Moved the compaction related ITs to a compaction package
