+++ test/src/main/java/org/apache/accumulo/test/compaction/CompactionExecutorIT.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.accumulo.test;
+package org.apache.accumulo.test.compaction;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
