+++ test/src/main/java/org/apache/accumulo/test/compaction/CompactionRateLimitingDeprecatedIT.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.test;
+package org.apache.accumulo.test.compaction;
 
 import org.apache.accumulo.core.conf.Property;
