+++ test/src/main/java/org/apache/accumulo/test/compaction/ExternalCompaction_3_IT.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.test;
+package org.apache.accumulo.test.compaction;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
