+++ test/src/main/java/org/apache/accumulo/test/compaction/TestCompactionCoordinator.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.test;
+package org.apache.accumulo.test.compaction;
 
 import java.io.IOException;
 import java.net.UnknownHostException;
